aggregation_layer/requirements.txt
paho-mqtt
PyYAML
torch